Don't log throwable when trying to run "hostname" binary
The code that tries to determine the hostname first tries to run the
"hostname" binary. This won't always be available on all platforms.
That's ok. We continue trying to check in another manner. Therefore
logging the exception will just create noise in the logs, it's not a
real issue.

Credits to @khiimstb for the original code